ship the compiled contracts in source distributions
this makes installation of the package a bit easier. solidity and the populus
python package don't have to be installed.

Strictly speaking this would not have been necessary for PyPI sdist uploads, but
I'd like to have a pure conscience.

There were some obstacles:
1. populus uses build as it's build directory. This is the same diretory that
python uses and we had to change python's build base to some other directory.
Otherwise sdist wouldn't include build/contracts.json.
2. We want to prevent any solidity compilation going on for sdist users. We do
this by checking for the file .compile-contracts.txt, which is excluded from the
sdist.