finally remove the old python trustlines-contracts package
trustlines-contracts-deploy in the py-deploy directory is the replacement. At
this point we could drop the '-deploy' suffix.

We could even think about moving it into the top-level folder, but that may be
confusing, since we really need to use 'make' at some point in order to install
it. At the moment I also fail to see a clear advantage.