Optimize gas costs
Optimize the gas costs by upgrading the solidity compiler to a
version
that can do better opimizations, and by seperation the values that need
to
be written during a hop from the ones that only needs to be read.

Additionally I refactor the naming of account to trustline, because
that
follows the naming conveention we no use elsewhere.