Fix fee calculation
Use proper rounding up formular
Use this formular in receiver pays functions and
reverse calculate the fees in the sender pays functions. This results in
the fees beeing the same in almost every case. Only exception is at
points where the fee calculation is not injective, because of the
rounding.
Closes #157