add a bit of infrastructure to test or call internal functions
This commit adds a TestCurrencyNetwork contract that subclasses the
CurrencyNetwork contract. It provides additional methods that can be used to
call into internal functions of CurrencyNetwork.

It should only be used for development and testing. It could be used to test
internal functions or export test data for the corresponding python
implementation of these internal functions.

The tldeploy.core.deploy_network functions takes an additional parameter
currency_network_contract_name. If it's not set, the function will use the
default name "CurrencyNetwork".

tl-deploy also takes a new --currency-network-contract-name. The test contract
can be deployed by running the following command:

    tl-deploy test --file addresses.json --currency-network-contract-name=TestCurrencyNetwork