enhance our Makefile
We've had problems with pip yesterday. make install wouldn't reinstall the
py-bin package. The reason was that pip version 9 was too old for this to work.

This is now solved by installing our requirements into the local virtualenv and
including pip>=18.1 in our requiremnts.

We're now able to run just make install in an empty virtulenv and get a working
environment.