export SETUPTOOLS_SCM_PRETEND_VERSION only where needed
This killed the build on travis in the venv-populus target, because some package
apparently used that variable, though it was only meant to be used by py-bin and
py-deploy. I'm not sure why this did only happen on travis at the moment.

We still set the variable when installing py-deploy and py-bin. For py-deploy
this could become problematic, because it has some dependencies. Let's hope for
the best!