avoid reassigning variables
When creating signed meta transactions, the code used to reassign values to
variables like in the following code:

     meta_transaction = MetaTransaction.from_function_call(function_call, to=to)
     meta_transaction = identity.filled_and_signed_meta_transaction(meta_transaction)

I've replaced that with a single assignment like:

     meta_transaction = identity.filled_and_signed_meta_transaction(
         MetaTransaction.from_function_call(function_call, to=to)
     )

At least for me this is a bit easier to read and avoids the reassignment.

Beware: the changed lines of code are formatted with black!