Make network expiration option optional for deploy
The expiration time option for the command to deploy a currency network
has a default value. Thereby it is not required to set expiration time
or -date anymore. The default is zero to disable the freezing of the
network. To disable it, one of the both expiration options must be set.

Related to: trustlines-network/project#682