Adjust behavior of currency deployment options
The expiration time does not have default value of zero anymore. It will
be only set to this default if the user has not set any of both
expiration related options. Thereby he can explicitly set it to zero
himself, but still get an error if the contradicting option is set as
well.

Related to: trustlines-network/project#682
See also: #256