Integrate recent features of deploy tools
This is especially focusing on the ability to specify transaction
options via CLI arguments. So does it allow to use a keystore to encrypt
raw transactions. Adding these options requires to update the whole flow
of the commands up to the deploy-tools.
Functions which are related to testing purpose only (not linked
directly or by reference to the CLI), have been leaved untouched by this
addition.

Moreover some other more convenient features have been integrated. As
example does the send_function_call_transaction already handle the
waiting for a transaction receipt.

Related to: #252