Add currency network adapter
Add a test adapter for the currency network to make the test easier to
understand and easier to maintain. Because the tests directly called the
smart contract code, a lot of tests were affected if something in
setAccount would change. this should not be the case anymore.
Some of the tests were already
adapted. Because of time constrains, this does not rewrite all tests.
Over time if touching test code we should no incremently adjust and use
the new adapter.