Add gas value snapshot testing
Add snapshot based testing for gas values.
The gas cost test are meant to notize sudden changes in gas costs and to
have a list of current gas costs and limits. It  should not be hard to
change the values when they changed for a valid reason. With this
change, the gas values are automatically stored in a file and can be
easily updated via a pytest flag.

I noticed that two of the tests sometimes change there gas costs, but I
will defer fixing this to later.