Change to revert when attempting zero tl update
Currently we were just returning without changing the trustline when an
attempt was made to change the trustlines from all zero to all zero.
This worked fine, but meant that you would not get an event and not a
failed transaction, which would mean no user feedback on what happend.
This now will instead revert, so giving a user feedback that this is not
possible.