Provide abi of all versions of currency networks merged together
tlbin now provides a merged abi package data file and a function
to load the abi of all versions of currency networks.

Pins the version of legacy currency networks abi.

This is useful for example for eth-index to figure out the abi
to decode events from currency network proxies that got upgraded