Trick gnosis safe to work with our debt tracking
Currently our identity calls the increaseDebt function when we call executeTransaction. This cannot work with gnosisSafe, but it appears that there is a workaround.
https://github.com/safe-global/safe-contracts/blob/da66b45ec87d2fb6da7dfd837b29eacdb9a604c5/contracts/common/SecuredTokenTransfer.sol

We could use a gasToken and increase the debt toward the delegate. Kind of hacky, but seems better than having to do a multi send for this.